refactor: code refactor for PR notes (DD-17)
