+++ src/app/features/admin/pages/tabs/route-tab/route-tab.component.html
@@ -10,7 +10,11 @@
 <tui-loader class="loader" [showLoader]="isLoading()">
   @for (trainRoute of routes(); track trainRoute.id) {
     <div tuiCardLarge tuiSurface="elevated" class="tui-space_top-4">
-      <dd-route-card [trainRoute]="trainRoute" (edit)="onEdit($event)" (delete)="onDelete($event)"></dd-route-card>
+      <dd-route-card
+        [trainRoute]="trainRoute"
+        (edit)="onEdit(trainRoute.id)"
+        (delete)="onDelete($event)"
+      ></dd-route-card>
     </div>
   }
 </tui-loader>
